fix: refactor ExportSettings component
